# frontend/lib/api/generated/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "DR-OS Control Plane",
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/projects": {
      "post": {
        "tags": [
          "projects"
        ],
        "summary": "Create Project",
        "operationId": "create_project_v1_projects_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProjectRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateProjectResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "projects"
        ],
        "summary": "List Projects",
        "operationId": "list_projects_v1_projects_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/projects/{project_id}": {
      "get": {
        "tags": [
          "projects"
        ],
        "summary": "Get Project",
        "operationId": "get_project_v1_projects__project_id__get",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Project Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectDetailResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "projects"
        ],
        "summary": "Update Project",
        "operationId": "update_project_v1_projects__project_id__patch",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Project Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProjectRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectDetailResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/projects/{project_id}/members": {
      "post": {
        "tags": [
          "projects"
        ],
        "summary": "Add Member",
        "operationId": "add_member_v1_projects__project_id__members_post",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Project Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddProjectMemberRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AddProjectMemberResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "projects"
        ],
        "summary": "List Members",
        "operationId": "list_members_v1_projects__project_id__members_get",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Project Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectMemberListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/projects/{project_id}/reviews": {
      "post": {
        "tags": [
          "reviews"
        ],
        "summary": "Create Review",
        "operationId": "create_review_v1_projects__project_id__reviews_post",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Project Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateReviewRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateReviewResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "reviews"
        ],
        "summary": "List Reviews",
        "operationId": "list_reviews_v1_projects__project_id__reviews_get",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Project Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/projects/{project_id}/reviews/{review_id}/decisions": {
      "post": {
        "tags": [
          "reviews"
        ],
        "summary": "Decide Review",
        "operationId": "decide_review_v1_projects__project_id__reviews__review_id__decisions_post",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Project Id"
            }
          },
          {
            "name": "review_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Review Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewDecisionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewDecisionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/projects/{project_id}/verify": {
      "post": {
        "tags": [
          "reviews"
        ],
        "summary": "Run Verification",
        "operationId": "run_verification_v1_projects__project_id__verify_post",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Project Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RunVerificationRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunVerificationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AddProjectMemberRequest": {
        "properties": {
          "principal_id": {
            "type": "string",
            "format": "uuid",
            "title": "Principal Id"
          },
          "role": {
            "$ref": "#/components/schemas/ProjectRole"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "principal_id",
          "role"
        ],
        "title": "AddProjectMemberRequest"
      },
      "AddProjectMemberResponse": {
        "properties": {
          "membership": {
            "$ref": "#/components/schemas/ProjectMemberRead"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "membership"
        ],
        "title": "AddProjectMemberResponse"
      },
...
      "ComplianceLevel": {
        "type": "string",
        "enum": [
          "public",
          "internal",
          "clinical"
        ],
        "title": "ComplianceLevel"
      },
...
      "CreateProjectRequest": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Name"
          },
          "project_type": {
            "$ref": "#/components/schemas/ProjectType"
          },
          "compliance_level": {
            "$ref": "#/components/schemas/ComplianceLevel",
            "default": "internal"
          },
          "owner_id": {
            "type": "string",
            "format": "uuid",
            "title": "Owner Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "name",
          "project_type",
          "owner_id"
        ],
        "title": "CreateProjectRequest"
      },
      "CreateProjectResponse": {
        "properties": {
          "project": {
            "$ref": "#/components/schemas/ProjectRead"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "project"
        ],
        "title": "CreateProjectResponse"
      },
      "CreateReviewRequest": {
        "properties": {
          "review_type": {
            "$ref": "#/components/schemas/ReviewType"
          },
          "target_kind": {
            "$ref": "#/components/schemas/LineageKind"
          },
          "target_id": {
            "type": "string",
            "format": "uuid",
            "title": "Target Id"
          },
          "reviewer_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reviewer Id"
          },
          "checklist_json": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Checklist Json"
          },
          "comments": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Comments"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "review_type",
          "target_kind",
          "target_id"
        ],
        "title": "CreateReviewRequest"
      },
      "CreateReviewResponse": {
        "properties": {
          "review": {
            "$ref": "#/components/schemas/ReviewRead"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "review"
        ],
        "title": "CreateReviewResponse"
      },
...
      "DatasetSnapshotRead": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "tenant_id": {
            "type": "string",
            "format": "uuid",
            "title": "Tenant Id"
          },
          "project_id": {
            "type": "string",
            "format": "uuid",
            "title": "Project Id"
          },
          "dataset_id": {
            "type": "string",
            "format": "uuid",
            "title": "Dataset Id"
          },
          "snapshot_no": {
            "type": "integer",
            "title": "Snapshot No"
          },
          "object_uri": {
            "type": "string",
            "title": "Object Uri"
          },
          "input_hash_sha256": {
            "type": "string",
            "title": "Input Hash Sha256"
          },
          "row_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Row Count"
          },
          "column_schema_json": {
            "additionalProperties": true,
            "type": "object",
            "title": "Column Schema Json"
          },
          "deid_status": {
            "$ref": "#/components/schemas/DeidStatus"
          },
          "phi_scan_status": {
            "$ref": "#/components/schemas/PhiScanStatus"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "tenant_id",
          "project_id",
          "dataset_id",
          "snapshot_no",
          "object_uri",
          "input_hash_sha256",
          "column_schema_json",
          "deid_status",
          "phi_scan_status",
          "created_at"
        ],
        "title": "DatasetSnapshotRead"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Page_ProjectRead_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ProjectRead"
            },
            "type": "array",
            "title": "Items"
          },
          "page": {
            "$ref": "#/components/schemas/PageInfo"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "items",
          "page"
        ],
        "title": "Page[ProjectRead]"
      },
...
      "ProjectDetailResponse": {
        "properties": {
          "project": {
            "$ref": "#/components/schemas/ProjectRead"
          },
          "active_workflows": {
            "items": {
              "$ref": "#/components/schemas/WorkflowInstanceRead"
            },
            "type": "array",
            "title": "Active Workflows"
          },
          "latest_snapshot": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DatasetSnapshotRead"
              },
              {
                "type": "null"
              }
            ]
          },
          "active_manuscript": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ManuscriptRead"
              },
              {
                "type": "null"
              }
            ]
          },
          "review_summary": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Review Summary"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "project"
        ],
        "title": "ProjectDetailResponse"
      },
      "ProjectListResponse": {
        "properties": {
          "items": {
            "$ref": "#/components/schemas/Page_ProjectRead_"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "items"
        ],
        "title": "ProjectListResponse"
      },
      "ProjectMemberListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ProjectMemberRead"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "items"
        ],
        "title": "ProjectMemberListResponse"
      },
...
      "ProjectRead": {
        "properties": {
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "tenant_id": {
            "type": "string",
            "format": "uuid",
            "title": "Tenant Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "project_type": {
            "$ref": "#/components/schemas/ProjectType"
          },
          "status": {
            "$ref": "#/components/schemas/ProjectState"
          },
          "compliance_level": {
            "$ref": "#/components/schemas/ComplianceLevel"
          },
          "owner_id": {
            "type": "string",
            "format": "uuid",
            "title": "Owner Id"
          },
          "active_manuscript_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Active Manuscript Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "created_at",
          "id",
          "tenant_id",
          "name",
          "project_type",
          "status",
          "compliance_level",
          "owner_id"
        ],
        "title": "ProjectRead"
      },
      "ProjectRole": {
        "type": "string",
        "enum": [
          "owner",
          "admin",
          "editor",
          "reviewer",
          "viewer"
        ],
        "title": "ProjectRole"
      },
...
      "ProjectType": {
        "type": "string",
        "enum": [
          "public_omics",
          "clinical_retrospective",
          "case_report",
          "grant"
        ],
        "title": "ProjectType"
      },
...
      "ReviewDecisionRequest": {
        "properties": {
          "action": {
            "type": "string",
            "minLength": 1,
            "title": "Action"
          },
          "comments": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Comments"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "action"
        ],
        "title": "ReviewDecisionRequest"
      },
      "ReviewDecisionResponse": {
        "properties": {
          "review": {
            "$ref": "#/components/schemas/ReviewRead"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "review"
        ],
        "title": "ReviewDecisionResponse"
      },
      "ReviewListResponse": {
        "properties": {
          "items": {
            "$ref": "#/components/schemas/Page_ReviewRead_"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "items"
        ],
        "title": "ReviewListResponse"
      },
...
      "RunVerificationRequest": {
        "properties": {
          "manuscript_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Manuscript Id"
          },
          "target_ids": {
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "type": "array",
            "title": "Target Ids"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "RunVerificationRequest"
      },
      "RunVerificationResponse": {
        "properties": {
          "workflow_instance_id": {
            "type": "string",
            "format": "uuid",
            "title": "Workflow Instance Id"
          },
          "gate_evaluations": {
            "items": {
              "$ref": "#/components/schemas/GateEvaluationRead"
            },
            "type": "array",
            "title": "Gate Evaluations"
          },
          "verifier_result": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/VerifierAgentResult"
              },
              {
                "type": "null"
              }
            ]
          },
          "blocking_summary": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Blocking Summary"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "workflow_instance_id"
        ],
        "title": "RunVerificationResponse"
      },
...
      "UpdateProjectRequest": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 200,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "status": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ProjectState"
              },
              {
                "type": "null"
              }
            ]
          },
          "compliance_level": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ComplianceLevel"
              },
              {
                "type": "null"
              }
            ]
          },
          "active_manuscript_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Active Manuscript Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "UpdateProjectRequest"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
      "WorkflowInstanceRead": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "tenant_id": {
            "type": "string",
            "format": "uuid",
            "title": "Tenant Id"
          },
          "project_id": {
            "type": "string",
            "format": "uuid",
            "title": "Project Id"
          },
          "workflow_type": {
            "type": "string",
            "title": "Workflow Type"
          },
          "state": {
            "$ref": "#/components/schemas/WorkflowState"
          },
          "current_step": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Step"
          },
          "parent_workflow_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Workflow Id"
          },
          "started_by": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started By"
          },
          "runtime_backend": {
            "$ref": "#/components/schemas/WorkflowBackend"
          },
          "started_at": {
            "type": "string",
            "format": "date-time",
            "title": "Started At"
          },
          "ended_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ended At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "tenant_id",
          "project_id",
          "workflow_type",
          "state",
          "runtime_backend",
          "started_at"
        ],
        "title": "WorkflowInstanceRead"
      },
...
    }
  }
}